test-suites/requirements.txt
optconstruct
pytest
docker
ansible
mock
reformat
autologging
odict
python-qpid-proton
